Updated Index.html
Shortened Title on line 11. Changed "divs" to "header", "nav", and fixed unnecessary spacing on lines 14 to lines 21. Closed the nav and header accordingly on lines 24 to lines 27. Wrapped hero in a section to follow proper HTML Semantic elements on line 28. changed the appropriate "divs" to "sections" and "articles" on lines 31-32, lines 34 -42, lines 50 - 52, lines 54-57, lines 61-62, lines 67-68, lines 73-74, and line 79. added footer and removed divs on line 83 and line 87